+++ apps/angular-lib-dev/src/app/app-routing.module.ts
@@ -37,11 +37,25 @@ const routes: Routes = [
     loadChildren: () =>
       import('./badge/badge.module').then((m) => m.BadgeModule),
   },
+  {
+    path: 'context-menu',
+    data: { name: 'Context Menu' },
+    loadChildren: () =>
+      import('./context-menu/context-menu.module').then(
+        (m) => m.ContextMenuModule
+      ),
+  },
+  {
+    path: 'modals',
+    data: { name: 'Modals' },
+    loadChildren: () =>
+      import('./modals/modals.module').then((m) => m.ModalsModule),
+  },
   { path: '**', redirectTo: 'buttons' },
 ]
 
 const routerOptions: ExtraOptions = {
-    anchorScrolling: 'enabled'
+  anchorScrolling: 'enabled',
 }
 
 @NgModule({