+++ apps/react-lib-dev/src/app/app.tsx
@@ -1,7 +1,6 @@
-import { Navbar, Tabs, Tab } from '@sebgroup/green-react'
-import { useState } from 'react'
-import { IValidator } from '@sebgroup/extract'
+import { Navbar, Tabs, Tab, AlertRibbon } from '@sebgroup/green-react'
 import { FormExample } from './form'
+import { AlertRibbons } from './alert-ribbons'
 
 export function App() {
   return (
@@ -17,8 +16,8 @@ export function App() {
                   <FormExample />
                 </div>
               </Tab>
-              <Tab title={'Other'}>
-                <div className="pt-5">Hello</div>
+              <Tab title={'Alert ribbons'}>
+                <AlertRibbons />
               </Tab>
             </Tabs>
           </div>