+++ libs/core/src/primitives/popover/popover.styles.ts
@@ -6,10 +6,6 @@ const style = css`
     background-color: white;
     box-shadow: 0 1rem 1rem 1rem rgba(0, 0, 0, 0.1);
   }
-
-  dialog::backdrop {
-    pointer-events: none;
-  }
 `
 
 export default style