+++ libs/react/src/types/props/index.ts
@@ -33,6 +33,14 @@ export interface SliderProps {
   onChange?: (value: number) => void
   enableClamping?: boolean
   onClamp?: (value: number) => void
+  enterKeyHint?:
+    | 'enter'
+    | 'done'
+    | 'go'
+    | 'next'
+    | 'previous'
+    | 'search'
+    | 'send'
 
   /**
    * @deprecated Use `value` instead